CLI+Meta: Reimplement CLI using click
Though argparse does have great features that can be utilized
underneath the surface, I found the click library to be much easier
to implement which will suit me well in my pursuit to push this hobby
project to focus on new features.

Refactoring has also taken place, mainly just trying to abide to the
demands of the linting software I was not aware of in the beginning
of project development.