Seperate file creation and update logic into seperate modules
With this seperation born concrete and abstract classes for both
the creator and updater modules. I questioned if the
behaviors should be created into classes but when thinking about
future needs for a configuration file I believe the abstractions
(as simple as they are) can be general enough for a config
file implementation to utilize.