File+Operations: Reimplement file operation classes as singletons
In an effort to escape the crutches of having global file
path variables, I decided to make changes to AbstractFileOperations
and in effect its subclasses. Why a singleton? In all honesty there
it doesn't have much of an effect, especially since we are not
using GUI elements that would see some form of event-driven
programming.

It was implemented in this way for two reasons: to ensure that the
data file path variable is unchangeable and for better readability;
also to allow testing of a temporary file to be possible.